refactor: dto -> entity, entity -> dto 변환 로직을 각 클래스가 갖도록 변경
